professors/requirements.txt
django==5.1.1
python-dotenv
django-cors-headers
psycopg2
djangorestframework
